requirements.txt
# DSL-SCHEMA-ProgramDesign-Transform 项目依赖
# 生成日期: 2026-02-14
# Python版本: >=3.9

# ========== Web框架 ==========
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# ========== 数据库 ==========
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pymongo>=4.6.0
pgvector>=0.2.4

# ========== AI/ML (可选依赖) ==========
# 安装这些依赖以启用完整功能
# torch>=2.1.0
# transformers>=4.36.0
# sentence-transformers>=2.2.2
# scikit-learn>=1.3.0
numpy>=1.24.0

# ========== LLM API ==========
openai>=1.6.0
anthropic>=0.8.0

# ========== HTTP客户端 ==========
requests>=2.31.0
httpx>=0.25.0

# ========== 工具库 ==========
PyYAML>=6.0.1
lark>=1.1.8
semver>=3.0.2
Pillow>=10.1.0

# ========== 可视化 ==========
matplotlib>=3.8.0
graphviz>=0.20.1

# ========== 测试 ==========
pytest>=7.4.0
pytest-asyncio>=0.21.0
